move more files
