+++ Sources/SmileID/Classes/Metadata/Device/AVCaptureDeviceExtension.swift
