+++ Sources/SmileID/Classes/Metadata/Device/BundleExtension.swift
